Move JsonPointerBuilder to own class
